+++ performance_test/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Removed an unused variable to fix a Clang build
+- Remove unused variable names in the `Plugin` abstract class
 
 ## 2.0.0 (2024/03/19)