+++ performance_test/CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Changed
 - Change the default `--history` arg from `KEEP_ALL` to `KEEP_LAST`
 - Change the default `--history-depth` arg from `1000` to `16`
+- If `--expected-num-pubs` is unspecified, set it to the same value as `-p`
+- If `--expected-num-subs` is unspecified, set it to the same value as `-s`
 ### Fixed
 - Removed an unused variable to fix a Clang build
 - Remove unused variable names in the `Plugin` abstract class